+++ src/graphql/resolvers/student_placement.js
@@ -78,19 +78,28 @@ module.exports={
             const ext = filename.substr(filename.lastIndexOf('.') + 1);
             const fileName = "StudentPlacement_"+getRegNo(req)+"_"+Placement_ID+"."+ext;
 
-            fs.unlinkSync(path.join(__dirname, "../../files/student-placements", fileName))
-            
+            if(fs.existsSync(path.join(__dirname, "../../files/student-placements", fileName))){
+                fs.unlinkSync(path.join(__dirname, "../../files/student-placements", fileName))
+            }
             await new Promise(res =>
                 createReadStream()
                 .pipe(fs.createWriteStream(path.join(__dirname, "../../files/student-placements", fileName)))
                 .on("close", res)
             );
-      
+            const Appointment_Order_Copy = path.join("files/student-placements", fileName);
+            await prisma.student_placement.update({
+                where:{
+                    Placement_ID
+                },
+                data:{
+                    Appointment_Order_Copy
+                }
+            })
             //files.push(filename);
             console.log(createReadStream)
 
 
-            return path.join("files/student-placements", fileName);
+            return Appointment_Order_Copy; 
         }
     }
